More minor fixes
- Made a single starter chest spawn in non-generous mode
- Tweaks to forge loot
- Revision to config code

+++ greymerk/roguelike/catacomb/Catacomb.java
@@ -86,7 +86,7 @@ public class Catacomb {
 			return false;
 		}
 		
-		if(!RogueConfig.getIntList(RogueConfig.DIMENSION).contains(world.provider.dimensionId)){
+		if(!RogueConfig.getIntList(RogueConfig.DIMENSION).contains((Integer)world.provider.dimensionId)){
 			return false;
 		}
